History does not display all entries - part 3
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/branches/0.6.5@6935 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ modules/history/history.h
@@ -37,7 +37,7 @@ struct HistoryEntry {
 
 struct HistoryDate {
 	QDateTime date;
-	uint idx;
+	quint64 idx;
 	HistoryDate();
 };
